New preview directive.
Adds a new directive called DIRECTIVE_PREVIEW with support for F_FILE type
observables in the gui. Modified the email analysis module to detect when a
plain text attachment is discovered and add this directive, so that plain text
emails are shown by default. Added a default configuration that specified how
many bytes to read out of the file for the preview.