Turned off VirusTotalURLAnalyzer. Needs to cache
I assumed this module was using the built-in VT cache but it
appears like it is not. I see where there are config items for
the cache, in the module, so maybe that didn't get completed yet?

This module destroyed our VT quota relatively quickly. If it is
using the cache, can it be improved or made more configurable?

There is a "quotas" field at api/v3/groups/{id} resource... maybe
we should have ace monitor that.