Created `analysis_module_graph_upn_analyzer`
This correlation module will look up the following information
on a user, given a valid UserPrincipleName:

    user_properties = [ 'id',
                        'employeeId',
                        'displayName',
                        'surname',
                        'givenName',
                        'city',
                        'state',
                        'streetAddress',
                        'country',
                        'postalCode',
                        'officeLocation',
                        'companyName',
                        'department',
                        'jobTitle',
                        'externalUserState',
                        'hireDate',
                        'lastPasswordChangeDateTime',
                        'mail',
                        'mobilePhone',
                        'onPremisesDistinguishedName',
                        'preferredLanguage',
                        'preferredName',
                        'usageLocation',
                        'userType',
                        'userPrincipalName']

Currently, the module requires you to specify a valid observable
to be treated like a UserPrincipleName. This is configurable and
the default is email_address:

    ; Required: Specify an existing observable type to treat like a uPn.
    ;  Often organizations use email addresses.
    upn_pointer = email_address

If an `employeeId` is returned then a F_USER observable is added to the analysis
with the value returned. This field is specified for unique user identifiers assigned
by the organization.

Additionally, if available, the following details on the user
will be added:

    - Azure AD Groups
    - Azure AD Directory Roles
    - Teams membersips
    - Direct Reports
    - Manager info
    - Profile photo

If a profile photo is obtained, it will be base64 encoded, stored in the analysis
details, and loaded in the GUI for the analyst.
The profile photo functionality configurable with the following two items:

    ; if true, get the profile photo for the user so it can be
    ; loaded with the analysis in the GUI.
    get_profile_photo = yes
    ; If get_profile_photo is true, download this size photo
    ; The supported sizes of HD photos on Microsoft 365 are as follows:
    ;   48x48, 64x64, 96x96, 120x120, 240x240, 360x360, 432x432, 504x504, and 648x648
    profile_photo_size = 240x240

If the `graph` integration is turned on, this module will be turned on and added to
the correlation module group.

This module uses the `GraphAnalysisModule`, which requires at least one
`graph_collection_account` to be configured.

Eventually, it may make sense to add a F_UPN observable.