Created `analysis_module_signin_history_analyzer`
This correlation module pulls the sign-in/auth history for the given
upn and provides some basic stats to the analyst.

The module requires you to specify a valid observable to be treated like
a UserPrincipleName, same as the `analysis_module_graph_upn_analyzer`.

The default query time is configurable at:

    ; How far back to query for the user's sign in activity
    day_interval = 14

Unique IPv4 addresses that authentication attempts were made from are
added as observables.

Hostname observables are added for unique devices identified as managed.