Added timeout to analysis modules with enforcement by the Worker.
In the event an analysis module does not return ACE will detect this condition and kill the analysis process. This timeout is set as a global option
[global]
maximum_analysis_time = seconds_until_death

and also at each analysis module
[analysis_module_whatever]
maximum_analysis_time = seconds_until_death

If the value does not exist in the analysis module then the global value is used instead.