Removed setting the gid for the ace user in the dev container.
When I realized that the local user ids and the ids used in the docker
container need to match up I changed it so that when the adduser is
executed to create the Docker container it is passed the current user's
UID and GID.

However, on a mac the GID is a super low number that is already used by
the Debain-based container. So I removed the GID setting.